+++ internal/decryptor/decryptor.go
@@ -34,6 +34,7 @@ import (
 	"github.com/getsops/sops/v3"
 	"github.com/getsops/sops/v3/aes"
 	"github.com/getsops/sops/v3/age"
+	"github.com/getsops/sops/v3/azkv"
 	"github.com/getsops/sops/v3/cmd/sops/common"
 	"github.com/getsops/sops/v3/cmd/sops/formats"
 	"github.com/getsops/sops/v3/keyservice"
@@ -50,7 +51,7 @@ import (
 
 	kustomizev1 "github.com/fluxcd/kustomize-controller/api/v1"
 	intawskms "github.com/fluxcd/kustomize-controller/internal/sops/awskms"
-	"github.com/fluxcd/kustomize-controller/internal/sops/azkv"
+	intazkv "github.com/fluxcd/kustomize-controller/internal/sops/azkv"
 	intkeyservice "github.com/fluxcd/kustomize-controller/internal/sops/keyservice"
 	"github.com/fluxcd/kustomize-controller/internal/sops/pgp"
 )
@@ -141,7 +142,7 @@ type Decryptor struct {
 	awsCredsProvider *awskms.CredentialsProvider
 	// azureToken is the Azure credential token used to authenticate towards
 	// any Azure Key Vault.
-	azureToken *azkv.Token
+	azureToken *azkv.TokenCredential
 	// gcpCredsJSON is the JSON credential file of the service account used to
 	// authenticate towards any GCP KMS.
 	gcpCredsJSON []byte
@@ -244,13 +245,15 @@ func (d *Decryptor) ImportKeys(ctx context.Context) error {
 			case filepath.Ext(DecryptionAzureAuthFile):
 				// Make sure we have the absolute name
 				if name == DecryptionAzureAuthFile {
-					conf := azkv.AADConfig{}
-					if err = azkv.LoadAADConfigFromBytes(value, &conf); err != nil {
+					conf := intazkv.AADConfig{}
+					if err = intazkv.LoadAADConfigFromBytes(value, &conf); err != nil {
 						return fmt.Errorf("failed to import '%s' data from %s decryption Secret '%s': %w", name, provider, secretName, err)
 					}
-					if d.azureToken, err = azkv.TokenFromAADConfig(conf); err != nil {
+					azureToken, err := intazkv.TokenCredentialFromAADConfig(conf)
+					if err != nil {
 						return fmt.Errorf("failed to import '%s' data from %s decryption Secret '%s': %w", name, provider, secretName, err)
 					}
+					d.azureToken = azkv.NewTokenCredential(azureToken)
 				}
 			case filepath.Ext(DecryptionGCPCredsFile):
 				if name == DecryptionGCPCredsFile {
@@ -554,19 +557,19 @@ func (d *Decryptor) sopsEncryptWithFormat(metadata sops.Metadata, data []byte, i
 }
 
 // keyServiceServer returns the SOPS (local) key service clients used to serve
-// decryption requests. loadKeyServiceServers() is only configured on the first
+// decryption requests. loadKeyServiceServer() is only configured on the first
 // call.
 func (d *Decryptor) keyServiceServer() []keyservice.KeyServiceClient {
 	d.localServiceOnce.Do(func() {
-		d.loadKeyServiceServers()
+		d.loadKeyServiceServer()
 	})
 	return d.keyServices
 }
 
-// loadKeyServiceServers loads the SOPS (local) key service clients used to
+// loadKeyServiceServer loads the SOPS (local) key service clients used to
 // serve decryption requests for the current set of Decryptor
 // credentials.
-func (d *Decryptor) loadKeyServiceServers() {
+func (d *Decryptor) loadKeyServiceServer() {
 	serverOpts := []intkeyservice.ServerOption{
 		intkeyservice.WithGnuPGHome(d.gnuPGHome),
 		intkeyservice.WithVaultToken(d.vaultToken),
