+++ internal/sops/azkv/config.go
@@ -1,16 +1,32 @@
-// Copyright (C) 2022 The Flux authors
-//
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+/*
+Copyright 2023 The Flux authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
 
 package azkv
 
 import (
+	"bytes"
+	"encoding/binary"
 	"fmt"
+	"io"
+	"unicode/utf16"
+
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/dimchansky/utfbom"
 	"sigs.k8s.io/yaml"
 )
 
@@ -49,7 +65,7 @@ type AZConfig struct {
 	Password string `json:"password,omitempty"`
 }
 
-// TokenFromAADConfig attempts to construct a Token using the AADConfig values.
+// TokenCredentialFromAADConfig attempts to construct a Token using the AADConfig values.
 // It detects credentials in the following order:
 //
 //   - azidentity.ClientSecretCredential when `tenantId`, `clientId` and
@@ -63,53 +79,40 @@ type AZConfig struct {
 //
 // If no set of credentials is found or the azcore.TokenCredential can not be
 // created, an error is returned.
-func TokenFromAADConfig(c AADConfig) (_ *Token, err error) {
-	var token azcore.TokenCredential
+func TokenCredentialFromAADConfig(c AADConfig) (token azcore.TokenCredential, err error) {
 	if c.TenantID != "" && c.ClientID != "" {
 		if c.ClientSecret != "" {
-			if token, err = azidentity.NewClientSecretCredential(c.TenantID, c.ClientID, c.ClientSecret, &azidentity.ClientSecretCredentialOptions{
+			return azidentity.NewClientSecretCredential(c.TenantID, c.ClientID, c.ClientSecret, &azidentity.ClientSecretCredentialOptions{
 				ClientOptions: azcore.ClientOptions{
 					Cloud: c.GetCloudConfig(),
 				},
-			}); err != nil {
-				return
-			}
-			return NewToken(token), nil
+			})
 		}
 		if c.ClientCertificate != "" {
 			certs, pk, err := azidentity.ParseCertificates([]byte(c.ClientCertificate), []byte(c.ClientCertificatePassword))
 			if err != nil {
 				return nil, err
 			}
-			if token, err = azidentity.NewClientCertificateCredential(c.TenantID, c.ClientID, certs, pk, &azidentity.ClientCertificateCredentialOptions{
+			return azidentity.NewClientCertificateCredential(c.TenantID, c.ClientID, certs, pk, &azidentity.ClientCertificateCredentialOptions{
 				SendCertificateChain: c.ClientCertificateSendChain,
 				ClientOptions: azcore.ClientOptions{
 					Cloud: c.GetCloudConfig(),
 				},
-			}); err != nil {
-				return nil, err
-			}
-			return NewToken(token), nil
+			})
 		}
 	}
 
 	switch {
 	case c.Tenant != "" && c.AppID != "" && c.Password != "":
-		if token, err = azidentity.NewClientSecretCredential(c.Tenant, c.AppID, c.Password, &azidentity.ClientSecretCredentialOptions{
+		return azidentity.NewClientSecretCredential(c.Tenant, c.AppID, c.Password, &azidentity.ClientSecretCredentialOptions{
 			ClientOptions: azcore.ClientOptions{
 				Cloud: c.GetCloudConfig(),
 			},
-		}); err != nil {
-			return
-		}
-		return NewToken(token), nil
+		})
 	case c.ClientID != "":
-		if token, err = azidentity.NewManagedIdentityCredential(&azidentity.ManagedIdentityCredentialOptions{
+		return azidentity.NewManagedIdentityCredential(&azidentity.ManagedIdentityCredentialOptions{
 			ID: azidentity.ClientID(c.ClientID),
-		}); err != nil {
-			return
-		}
-		return NewToken(token), nil
+		})
 	default:
 		return nil, fmt.Errorf("invalid data: requires a '%s' field, a combination of '%s', '%s' and '%s', or '%s', '%s' and '%s'",
 			"clientId", "tenantId", "clientId", "clientSecret", "tenantId", "clientId", "clientCertificate")
@@ -126,4 +129,25 @@ func (s AADConfig) GetCloudConfig() cloud.Configuration {
 		}
 	}
 	return cloud.AzurePublic
+}
+
+func decode(b []byte) ([]byte, error) {
+	reader, enc := utfbom.Skip(bytes.NewReader(b))
+	switch enc {
+	case utfbom.UTF16LittleEndian:
+		u16 := make([]uint16, (len(b)/2)-1)
+		err := binary.Read(reader, binary.LittleEndian, &u16)
+		if err != nil {
+			return nil, err
+		}
+		return []byte(string(utf16.Decode(u16))), nil
+	case utfbom.UTF16BigEndian:
+		u16 := make([]uint16, (len(b)/2)-1)
+		err := binary.Read(reader, binary.BigEndian, &u16)
+		if err != nil {
+			return nil, err
+		}
+		return []byte(string(utf16.Decode(u16))), nil
+	}
+	return io.ReadAll(reader)
 }
