+++ internal/sops/keyservice/options.go
@@ -9,10 +9,10 @@ package keyservice
 import (
 	extage "filippo.io/age"
 	"github.com/getsops/sops/v3/age"
+	"github.com/getsops/sops/v3/azkv"
 	"github.com/getsops/sops/v3/keyservice"
 	awskms "github.com/getsops/sops/v3/kms"
 
-	"github.com/fluxcd/kustomize-controller/internal/sops/azkv"
 	"github.com/fluxcd/kustomize-controller/internal/sops/gcpkms"
 	"github.com/fluxcd/kustomize-controller/internal/sops/hcvault"
 	"github.com/fluxcd/kustomize-controller/internal/sops/pgp"
@@ -69,7 +69,7 @@ func (o WithGCPCredsJSON) ApplyToServer(s *Server) {
 
 // WithAzureToken configures the Azure credential token on the Server.
 type WithAzureToken struct {
-	Token *azkv.Token
+	Token *azkv.TokenCredential
 }
 
 // ApplyToServer applies this configuration to the given Server.
