+++ internal/sops/keyservice/utils_test.go
@@ -11,11 +11,11 @@ import (
 	"fmt"
 
 	"github.com/getsops/sops/v3/age"
+	"github.com/getsops/sops/v3/azkv"
 	"github.com/getsops/sops/v3/keys"
 	"github.com/getsops/sops/v3/keyservice"
 	awskms "github.com/getsops/sops/v3/kms"
 
-	"github.com/fluxcd/kustomize-controller/internal/sops/azkv"
 	"github.com/fluxcd/kustomize-controller/internal/sops/gcpkms"
 	"github.com/fluxcd/kustomize-controller/internal/sops/hcvault"
 	"github.com/fluxcd/kustomize-controller/internal/sops/pgp"
